// Downloads/dpm-ui-web-and-mobile-fixes 1/dpm-ui-web-and-mobile-fixes/web-apps/consumer-portal/src/types/quoteAndBuy.ts
  interface TravelTaxFeeBreakdown {
    amount: number;
    percentage: number | null;
    type: string;
  }
  
  interface TravelPricingOption {
    finalAmount: number;
    premiumDue: number;
    taxFeeBreakdowns: TravelTaxFeeBreakdown[];
  }
  
  interface TravelPremiumInfo {
    annualPremium: number;
    finalPremium: number;
    grossPremium: number;
    netPremium: number;
  }
  
  interface TravelPurchasedCoverage {
    coverageCode: string;
    coverageName: string;
    premiumInfo: TravelPremiumInfo;
  }
  
  interface TravelAPIModel {
    policyEffectiveDate: string;
    policyExpiryDate: string;
    pricingOptions: TravelPricingOption[];
    purchasedCoverage: TravelPurchasedCoverage[];
    quotationDate: string;
    renewalInd: string;
  }
  
  interface TravelAPIData {
    model: TravelAPIModel;
    status: string;
  }
  
  export interface TravelApiResponse {
    pricingOptions: TravelPricingOption[];
    purchasedCoverage: any;
    code: number;
    message: string;
    transactionId: string | null;
    customerRefNum: string | null;
    data: TravelAPIData;
    errors: any;
  }